+++ yggdrash-validator/src/main/java/io/yggdrash/validator/GrpcNodeServer.java
@@ -3,18 +3,13 @@ package io.yggdrash.validator;
 import com.google.gson.Gson;
 import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
-import io.grpc.stub.StreamObserver;
-import io.yggdrash.common.config.DefaultConfig;
 import io.yggdrash.core.blockchain.Block;
 import io.yggdrash.core.exception.NotValidateException;
 import io.yggdrash.core.wallet.Wallet;
-import io.yggdrash.proto.ConsensusEbftGrpc;
-import io.yggdrash.proto.EbftProto;
 import io.yggdrash.validator.data.BlockCon;
 import io.yggdrash.validator.data.BlockConChain;
 import io.yggdrash.validator.data.NodeStatus;
 import io.yggdrash.validator.util.TestUtils;
-import org.lognet.springboot.grpc.GRpcService;
 import org.slf4j.LoggerFactory;
 import org.spongycastle.util.encoders.Hex;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -22,13 +17,13 @@ import org.springframework.boot.CommandLineRunner;
 import org.springframework.core.io.ClassPathResource;
 import org.springframework.scheduling.annotation.EnableScheduling;
 import org.springframework.scheduling.annotation.Scheduled;
+import org.springframework.stereotype.Service;
 import org.springframework.util.FileCopyUtils;
 
 import java.io.IOException;
 import java.net.InetAddress;
 import java.nio.charset.StandardCharsets;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
@@ -37,38 +32,37 @@ import java.util.concurrent.locks.ReentrantLock;
 
 import static io.yggdrash.common.util.Utils.sleep;
 
-@GRpcService
+@Service
 @EnableScheduling
-public class GrpcNodeServer extends ConsensusEbftGrpc.ConsensusEbftImplBase
-        implements CommandLineRunner {
+public class GrpcNodeServer implements CommandLineRunner {
 
     private static final org.slf4j.Logger log = LoggerFactory.getLogger(GrpcNodeServer.class);
 
-    private static final boolean TEST_OMIT_VERIFY = false;
     private static final boolean ABNORMAL_TEST = false;
 
-    private final int consenusCount;
     private final boolean isValidator;
+    private final int consenusCount;
+
     private final Wallet wallet;
     private final BlockConChain blockConChain;
+
     private final GrpcNodeClient myNode;
     private final Map<String, GrpcNodeClient> totalValidatorMap;
 
     private boolean isActive;
     private boolean isSynced;
 
-    private DefaultConfig defaultConfig;
     private ReentrantLock lock = new ReentrantLock();
 
     @Autowired
-    public GrpcNodeServer(DefaultConfig defaultConfig, Wallet wallet, BlockConChain blockConChain) {
-        this.defaultConfig = defaultConfig;
+    public GrpcNodeServer(Wallet wallet, BlockConChain blockConChain) {
         this.wallet = wallet;
         this.blockConChain = blockConChain;
         this.myNode = initMyNode();
         this.totalValidatorMap = initTotalValidator();
         this.isValidator = initValidator();
         this.isActive = false;
+        this.isSynced = false;
         if (totalValidatorMap != null) {
             this.consenusCount = totalValidatorMap.size() / 2 + 1;
         } else {
@@ -111,83 +105,6 @@ public class GrpcNodeServer extends ConsensusEbftGrpc.ConsensusEbftImplBase
 
         loggingNode();
 
-    }
-
-    @Override
-    public void pingPongTime(EbftProto.PingTime request,
-                             StreamObserver<EbftProto.PongTime> responseObserver) {
-        long timestamp = System.currentTimeMillis();
-        EbftProto.PongTime pongTime
-                = EbftProto.PongTime.newBuilder().setTimestamp(timestamp).build();
-        responseObserver.onNext(pongTime);
-        responseObserver.onCompleted();
-    }
-
-    @Override
-    public void getNodeStatus(
-            EbftProto.Chain request,
-            StreamObserver<io.yggdrash.proto.EbftProto.NodeStatus> responseObserver) {
-        NodeStatus newNodeStatus = getMyNodeStatus();
-        responseObserver.onNext(NodeStatus.toProto(newNodeStatus));
-        responseObserver.onCompleted();
-    }
-
-    @Override
-    public void exchangeNodeStatus(io.yggdrash.proto.EbftProto.NodeStatus request,
-            io.grpc.stub.StreamObserver<io.yggdrash.proto.EbftProto.NodeStatus> responseObserver) {
-        NodeStatus blockStatus = new NodeStatus(request);
-        updateStatus(blockStatus);
-
-        NodeStatus newNodeStatus = getMyNodeStatus();
-        responseObserver.onNext(NodeStatus.toProto(newNodeStatus));
-        responseObserver.onCompleted();
-    }
-
-    @Override
-    public void broadcastBlockCon(io.yggdrash.proto.EbftProto.BlockCon request,
-            io.grpc.stub.StreamObserver<io.yggdrash.proto.NetProto.Empty> responseObserver) {
-        BlockCon newBlockCon = new BlockCon(request);
-        if (!BlockCon.verify(newBlockCon) || !consensusVerify(newBlockCon)) {
-            log.error("Verify Fail");
-            return;
-        }
-
-        BlockCon lastBlockCon = this.blockConChain.getLastConfirmedBlockCon();
-
-        responseObserver.onNext(io.yggdrash.proto.NetProto.Empty.newBuilder().build());
-        responseObserver.onCompleted();
-
-        if (lastBlockCon.getIndex() == newBlockCon.getIndex() - 1
-                && Arrays.equals(lastBlockCon.getHash(), newBlockCon.getPrevBlockHash())) {
-
-            lock.lock();
-            updateUnconfirmedBlock(newBlockCon);
-            lock.unlock();
-        }
-    }
-
-    @Override
-    public void getBlockConList(io.yggdrash.proto.EbftProto.Offset request,
-            io.grpc.stub.StreamObserver<EbftProto.BlockConList> responseObserver) {
-        long start = request.getIndex();
-        long count = request.getCount();
-        List<BlockCon> blockConList = new ArrayList<>();
-
-        long end = Math.min(start - 1 + count,
-                this.blockConChain.getLastConfirmedBlockCon().getIndex());
-
-        log.debug("start: " + start);
-        log.debug("end: " + end);
-
-        if (start < end) {
-            for (long l = start; l <= end; l++) {
-                blockConList.add(this.blockConChain.getBlockConStore().get(
-                        this.blockConChain.getBlockConKeyStore().get(l)));
-            }
-        }
-
-        responseObserver.onNext(BlockCon.toProtoList(blockConList));
-        responseObserver.onCompleted();
     }
 
     private void checkNode() {
@@ -240,18 +157,6 @@ public class GrpcNodeServer extends ConsensusEbftGrpc.ConsensusEbftImplBase
             }
         } else {
             client.setIsRunning(false);
-        }
-    }
-
-    private void updateStatus(NodeStatus nodeStatus) {
-        if (NodeStatus.verify(nodeStatus)) {
-            for (BlockCon blockCon : nodeStatus.getUnConfirmedBlockConList()) {
-                if (blockCon.getIndex()
-                        <= this.blockConChain.getLastConfirmedBlockCon().getIndex()) {
-                    continue;
-                }
-                updateUnconfirmedBlock(blockCon);
-            }
         }
     }
 
@@ -541,7 +446,7 @@ public class GrpcNodeServer extends ConsensusEbftGrpc.ConsensusEbftImplBase
         }
     }
 
-    private void updateUnconfirmedBlock(BlockCon blockCon) {
+    public void updateUnconfirmedBlock(BlockCon blockCon) {
         if (this.blockConChain.getUnConfirmedBlockConMap().containsKey(blockCon.getHashHex())) {
             // if exist, update consensus
             if (blockCon.getConsensusList().size() > 0) {
@@ -559,7 +464,7 @@ public class GrpcNodeServer extends ConsensusEbftGrpc.ConsensusEbftImplBase
         }
     }
 
-    private NodeStatus getMyNodeStatus() {
+    public NodeStatus getMyNodeStatus() {
         NodeStatus newNodeStatus = new NodeStatus(this.getActiveNodeList(),
                 this.blockConChain.getLastConfirmedBlockCon(),
                 new ArrayList<>(this.blockConChain.getUnConfirmedBlockConMap().values()));
@@ -606,11 +511,11 @@ public class GrpcNodeServer extends ConsensusEbftGrpc.ConsensusEbftImplBase
     }
 
     private GrpcNodeClient initMyNode() {
-        byte[] realPubKey = new byte[64];
-        System.arraycopy(wallet.getPubicKey(), 1, realPubKey, 0, 64);
-        GrpcNodeClient client = new GrpcNodeClient(Hex.toHexString(realPubKey),
+        GrpcNodeClient client = new GrpcNodeClient(
+                wallet.getPubicKeyHex().substring(2),
                 InetAddress.getLoopbackAddress().getHostAddress(),
                 Integer.parseInt(System.getProperty("grpc.port")));
+
         client.setMyclient(true);
         client.setIsRunning(true);
 
@@ -663,11 +568,7 @@ public class GrpcNodeServer extends ConsensusEbftGrpc.ConsensusEbftImplBase
         return count;
     }
 
-    private boolean consensusVerify(BlockCon blockCon) {
-        if (TEST_OMIT_VERIFY) {
-            return true;
-        }
-
+    public boolean consensusVerify(BlockCon blockCon) {
         if (blockCon.getConsensusList().size() <= 0) {
             return true;
         }
@@ -682,29 +583,8 @@ public class GrpcNodeServer extends ConsensusEbftGrpc.ConsensusEbftImplBase
         return true;
     }
 
-    private int getUnconfirmedConsenusCount() {
-        int count = 0;
-        long index = this.blockConChain.getLastConfirmedBlockCon().getIndex() + 1;
-        for (String key : this.blockConChain.getUnConfirmedBlockConMap().keySet()) {
-            BlockCon blockCon = this.blockConChain.getUnConfirmedBlockConMap().get(key);
-            if (count < blockCon.getConsensusList().size()
-                    && blockCon.getIndex() == index) {
-                count = blockCon.getConsensusList().size();
-            }
-        }
-        return count;
-    }
-
-    private void loggingMap(Map<String, BlockCon> map) {
-        log.debug("[" + map.size() + "]");
-        for (String key : map.keySet()) {
-            BlockCon blockCon = map.get(key);
-            log.debug(blockCon.getBlock().getAddressHexString()
-                    + " "
-                    + "["
-                    + blockCon.getIndex()
-                    + "]"
-                    + blockCon.getHashHex());
-        }
+    // todo: check security
+    public ReentrantLock getLock() {
+        return lock;
     }
 }
